% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.conv.qf.not.vrbs.R
\name{def.conv.qf.not.vrbs}
\alias{def.conv.qf.not.vrbs}
\title{Definition function: Convert from verbose to non-verbose output of quality tests}
\usage{
def.conv.qf.not.vrbs(qf)
}
\arguments{
\item{qf}{A list of variables, each containing a data frame of quality flags for that variable.}
}
\value{
A named list of variables matching those in \code{qf}, each itself a named list of flags 
(standard or user-defined names), with nested lists of $fail and $na vector positions of failed and na 
quality test results for that variable and flag (eg. posQf$X$posQfStep$fail and posQf$Y$posQfStep$na).
}
\description{
Function definition. Convert from the verbose output of quality test algorithms (actual quality flag values) to the non-verbose option (vector positions of failed and na test results).
}
\examples{
qf <- list(var01=data.frame(qfRng = c(0,1,-1,0,1,0,0,0,0,-1),qfStep = c(0,0,-1,-1,0,0,0,0,0,-1))) # Verbose output from quality testing
posQf <- def.conv.qf.not.vrbs(qf=qf) # Convert to vector positions of failed and na quality tests
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\code{\link[eddy4R.qaqc]{def.plau}}
\code{\link[eddy4R.qaqc]{def.conv.qf.vrbs}}
}
\author{
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\keyword{flag}
\keyword{quality}
