% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dspk.br86.R
\name{def.dspk.br86}
\alias{def.dspk.br86}
\title{Definition function: Median filter de-spiking}
\usage{
def.dspk.br86(dataInp, WndwFilt = 9, NumBin = 2, ThshReso = 10,
  FracRealMin = 2.5e-02)
}
\arguments{
\item{dataInp}{Required. A univariate vector of integers or numerics of Input data}

\item{WndwFilt}{Optional. A single integer value of filter width. Default = 9}

\item{NumBin}{Optional. A single integer value of the initial number/step size of histogram bins. Default = 2}

\item{ThshReso}{Optional. A single integer value of the resolution threshold for spike determination. Only considered spike only if difference larger than measurement resolution x ThshReso. Default = 10}

\item{FracRealMin}{Optional. A single numeric value of the minimum fraction of non-NA values required to perform median filter despiking. Default = 0.025 (2.5\%)}
}
\value{
A list of:

\code{dataInp} Same as input.

\code{WndwFilt} Same as input.

\code{NumBin} Same as input.

\code{ThshReso} Same as input.

\code{FracRealMin} Same as input.

\code{numBinFinl} Integer. The final number of histogram bins used.

\code{numSpk} Integer. The number of spikes identified.

\code{dataOut} Numeric vector of input data with spikes removed.

\code{dataNorm} Numeric vector of input data normalized to mean 0 and standard deviation of 1

\code{dataNormDiff} Numeric vector of differences between subsequent values of \code{dataNorm}. Length = 1-length(dataInp)

\code{resoDataNorm} Numeric value. Measurement resolution (assumed to be smallest recorded change in \code{dataNorm})

\code{thshNumData} Numeric value. Minimum number of non-NAs in window to calculate median, otherwise NA is returned.

\code{dataNormFiltMed} Numeric vector. Median-filtered timeseries of \code{dataNorm}.

\code{histDiff} Numeric vector. Initial histogram of differences of \code{dataNormFiltMed}

\code{crit} Logical. Criteria for ending iteration over histogram bins

\code{locBin} Numeric vector. Bin edges for \code{histDiff}

\code{histDiffFinl} Numeric vector. Final histogram of differences of \code{dataNormFiltMed} using bin edges in \code{locBin}

\code{histDiffFinl$counts} Integer vector. Counts within each bin of \code{histDiffFinl}

\code{histDiffFinl$breaks} Numeric vector. Bin edges for \code{histDiffFinl}

\code{idxBinMin} Integer. Index of histogram bin within \code{histDiffFinl} with minimum number of counts.

\code{idxBinMax} Integer. Index of histogram bin within \code{histDiffFinl} with maximum number of counts.

\code{idxThshBinMin} Integer. Current iteration of minimum threshold index of histogram bin within \code{histDiffFinl} for spike determination

\code{idxThshBinMax}  Integer. Current iteration of maximum threshold index of histogram bin within \code{histDiffFinl} for spike determination

\code{idxThshBinMinFinl} Integer. Final minimum threshold index of histogram bin within \code{histDiffFinl} for spike determination

\code{idxThshBinMaxFinl} Integer. Final maximum threshold index of histogram bin within \code{histDiffFinl} for spike determination

\code{setSpk} Indices of determined spikes within \code{dataInp}

\code{qfSpk} Integer. Quality flag values [-1,0,1] for despike test, matching same size as \code{dataInp}
}
\description{
Function definition. Median filter de-spiking\cr
After Brock (1986), Starkenburg et al. (2014); \cr
Order N = 3 - 4 (i.e., a window of 7 - 9 points); \cr
Such a filter will be sensitive to spikes of up to 3 - 4 consecutive points (Brock, 1986);
this is the maximum number of points typically allowed to be considered spikes (Vickers and Mahrt, 1997; Mauder and Foken, 2011);
Raw signal is normalized so that its mean is zero and standard deviation is 1; \cr
Initial number of bins is 2, and doubles iteratively until minima are found; \cr
Distribution  of the differences (D1) between the filtered signal and the raw signal is assessed;
good data points will be centered within the histogram near zero;
spikes will lie farthest from the center, resulting in subpopulations that make the distribution multi-modal; \cr
Determine the the range of accepted values, DT by searching the histogram for the first minima from the center;
Points where |D1| > DT are considered spikes, provided the difference is >= 10 x the measurement resolution (smallest detected change); \cr
Error escapes:\cr
-1: no non-NAs in dataset \cr
-2: singular or constant value \cr
-3: measurement changes so slow that time-series and filter value are identical \cr
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
Brock, F. V. A Nonlinear Filter to Remove Impulse Noise from Meteorological Data. J. Atmos. Oceanic Technol. 3, 51–58 (1986). \cr
Starkenburg, D. et al. Assessment of Despiking Methods for Turbulence Data in Micrometeorology. J. Atmos. Oceanic Technol. 33, 2001–2013 (2016). \cr
Vickers, D. & Mahrt, L. Quality control and flux sampling problems for tower and aircraft data. in 512–526 (Amer Meteorological Soc, 1997). \cr
Mauder, M. and Foken, T. Documentation and instruction manual of the edy covariance software package TK3. Arbeitsergebn. Univ Bayreuth. Abt Mikrometeorol. ISSN 1614-8916, 46:58 pp. (2011)
NEON Algorithm Theoretical Basis Document: Quality Flags and Quality Metrics for TIS Data Products (NEON.DOC.001113) \cr
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\keyword{Currently}
\keyword{none}
