% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dspk.wndw.R
\name{def.dspk.wndw}
\alias{def.dspk.wndw}
\title{Determine spike locations using window-based statistics}
\usage{
def.dspk.wndw(data, Trt = list(AlgClss = c("mean", "median")[2], NumPtsWndw =
  c(10, 101)[2], NumPtsSlid = 1, ThshStd = c(3.5, 20)[2], NaPropMax = 0.1, Infl
  = 0, IterMax = Inf, NumPtsGrp = c(4, 10)[2], NaTrt = c("approx", "omit")[2]),
  Cntl = list(NaOmit = c(TRUE, FALSE)[2], Prnt = c(TRUE, FALSE)[1], Plot =
  c(TRUE, FALSE)[1]))
}
\arguments{
\item{data}{Required input. A data frame or matrix containing the data to be evaluated.}

\item{Trt}{Optional. A list of the following parameters specifying the details of the despike algorithm. \cr
AlgClss: string. de-spiking algorithm class ["mean" or "median"] \cr
NumPtsWndw: integer. window size [data points] (must be odd for median / mad) \cr
NumPtsSlid: integer. window sliding increment [data points] \cr
ThshStd: number. threshold for detecting data point as spike [sigma / MAD_sigma] \cr
NaPropMax: The maximum allowable proportion of NA values per window in which spikes can be reliably determined \cr
Infl: number. inflation per iteration [fraction of sigma / MAD_sigma] \cr
IterMax: number or Inf. maximum number of iterations \cr
NumPtsGrp. integer. minimum group size that is not considered as consecutive spikes [data points] \cr
NaTrt. string. spike handling among iterations ["approx" or "omit"]}

\item{Cntl}{Optional. A list of the following parameters specifying other implementation details \cr
NaOmit: Logical. delete leading / trailing NAs from dataset? \cr
Prnt: Logical. print results? \cr
Plot: Logical. plot results?}
}
\value{
A list of the following: \cr
data: the despiked data matrix \cr
smmy: a summary of the despike algorithm results, including iterations (iter), determined spikes (news), and total resultant NAs (alls) \cr
posSpk: a list consisting of $fail and $na, each of length equal to the number of data variables giving the failed positions of determined spikes and unable-to-evaluate positions, respectively
}
\description{
Function definition. Determines spike locations based on window-based Gaussian statistics (arithmetic mean and standard deviation) and distribution statistics (median, median absolute deviation).
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Cove Sturtevant \email{csturtevant@neoninc.org}
}
\references{
Hojstrup, J.: A statistical data screening procedure, Meas. Sci. Technol., 4, 153-157, doi:10.1088/0957-0233/4/2/003, 1993. \cr
Metzger, S., Junkermann, W., Mauder, M., Beyrich, F., Butterbach-Bahl, K., Schmid, H. P., and Foken, T.: Eddy-covariance flux measurements with a weight-shift microlight aircraft, Atmos. Meas. Tech., 5, 1699-1717, doi:10.5194/amt-5-1699-2012, 2012. \cr
Vickers, D., and Mahrt, L.: Quality control and flux sampling problems for tower and aircraft data, J. Atmos. Oceanic Technol., 14, 512-526, doi:10.1175/1520-0426(1997)014<0512:QCAFSP>2.0.CO;2, 1997. \cr
}
\seealso{
Currently none
}
\keyword{de-spiking}
\keyword{spike,}

