% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.mapp.fulc.form.key.R
\name{def.mapp.fulc.form.key}
\alias{def.mapp.fulc.form.key}
\title{Map data names to internal keys in Fulcrum form schema}
\usage{
def.mapp.fulc.form.key(schmForm)
}
\arguments{
\item{\code{schmForm}}{Required. The Fulcrum form schema as extracted from the Fulcrum API GET 
request (using the use httr::content function, see examples)}
}
\value{
A data frame of \code{name}, \code{key}, and \code{path} mappings. \cr
}
\description{
Function definition. Parses Fulcrum form schema to map all data names to internal keys
}
\examples{
urlApiForm <- paste0("https://api.fulcrumapp.com/api/v2/forms/",idForm,".json") # URL for specific form id, where idForm is a string with your form id
rspnForm <- httr::GET(urlApiForm,add_headers("X-ApiToken" = keyApi, Accept = "application/json")) # where keyApi is a string with your API key
schmForm <- httr::content(rspnForm) # The overall form schema
fldMapp <- def.mapp.fulc.form.key(schmForm) # Grab the mapping between the field names and keys
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org}
}
\references{
license: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{Fulcrum}

