% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.neon.ecte.dp01.qf.grup.R
\name{def.neon.ecte.dp01.qf.grup}
\alias{def.neon.ecte.dp01.qf.grup}
\title{Definition function: Grouping the quality flags for each of NEON eddy-covariance turbulent exchange L1 data product}
\usage{
def.neon.ecte.dp01.qf.grup(qfSens00, qfSens01 = NULL, qfSens02 = NULL,
  dp01 = c("irgaCo2", "irgaH2o", "soni", "soniAmrs")[1])
}
\arguments{
\item{\code{qfSens00}}{A dataframe containing the input flag data that related to L1 data products are being grouped. Of class integer". [-]}

\item{\code{qfSens01}}{Optional. A dataframe containing the input flag data that related to L1 data products are being grouped. Of class integer". [-]}

\item{\code{qfSens02}}{Optional. A dataframe containing the input flag data that related to L1 data products are being grouped. Of class integer". [-]}

\item{\code{dp01}}{A vector of class "character" containing the name of NEON eddy-covariance turbulent exchange L1 data products which the flags are being grouped, c("irgaCo2","irgaH2o","soni","soniAmrs"). Defaults to "irgaCo2". [-]}
}
\value{
A list of data frame of the quality flags related to that sub-data product. \cr
}
\description{
Function definition. Grouping the quality flags of each NEON eddy-covariance turbulent exchange L1 data product into a single dataframe for further use in the calculation of Alpha, Beta, and Final flag.
}
\examples{
#generate the fake quality flags for each sensor
TimeBgn <- "2016-04-24 02:00:00.000"
TimeEnd <- "2016-04-24 02:29:59.950"
inp <- list()
inp$qf$irga <- eddy4R.qaqc::def.qf.ecte(TimeBgn = TimeBgn, TimeEnd = TimeEnd, Freq = 20, Sens = "irga", PcntQf = 0.05)
inp$qf$irgaMfcSamp <- eddy4R.qaqc::def.qf.ecte(TimeBgn = TimeBgn, TimeEnd = TimeEnd, Freq = 20, Sens = "irgaMfcSamp", PcntQf = 0.05)
inp$qf$soni <- eddy4R.qaqc::def.qf.ecte(TimeBgn = TimeBgn, TimeEnd = TimeEnd, Freq = 20, Sens = "soni", PcntQf = 0.05)
inp$qf$soniAmrs <- eddy4R.qaqc::def.qf.ecte(TimeBgn = TimeBgn, TimeEnd = TimeEnd, Freq = 40, Sens = "soniAmrs", PcntQf = 0.05)

#grouping the set of the flags
qfGrupIrgaCo2 <- def.neon.ecte.dp01.qf.grup(qfSens00 = inp$qf$irga, qfSens01 = inp$qf$irgaMfcSamp, dp01="irgaCo2")
qfGrupSoni <- def.neon.ecte.dp01.qf.grup(qfSens00 = inp$qf$soni, qfSens01 = inp$qf$irga, dp01="soni")
}
\author{
Natchaya Pingintha-Durden \email{ndurden@neoninc.org} \cr
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{NEON}
\keyword{QAQC,}
\keyword{flags,}
\keyword{metrics}
\keyword{quality}

