% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.plau.R
\name{def.plau}
\alias{def.plau}
\title{Plausibility tests (Range, Step, Persistence, Null, Gap)}
\usage{
def.plau(data, ts = as.POSIXlt(seq.POSIXt(from = Sys.time(), by = "sec",
  length.out = length(data[, 1]))), RngMin = apply(data, 2, min, na.rm =
  TRUE), RngMax = apply(data, 2, max, na.rm = TRUE),
  DiffStepMax = apply(abs(apply(data, 2, diff)), 2, max, na.rm = TRUE),
  DiffPersMin = rep.int(0, length(data)), TintPers = 60 *
  median(abs(diff(ts)), na.rm = TRUE) * rep.int(1, length(data)),
  TestNull = rep(FALSE, length(data)), NumGap = length(data[, 1]) + 1)
}
\arguments{
\item{\code{data}}{Required input. A data frame containing the data to be evaluated (do not include the time stamp vector here).}

\item{\code{ts}}{Optional. A time vector of class POSIXlt of times corresponding with each row in data. Defaults to an evenly spaced time vector starting from system time of execution by seconds.}

\item{\code{RngMin}}{Optional. A numeric vector of length equal to number of variables in data containing the minimum acceptable value for each variable. Defaults to observed minimums (no flags will result)}

\item{\code{RngMax}}{Optional. A numeric vector of length equal to number of variables in data containing the maximum acceptable value for each variable. Defaults to observed maximums (no flags will result)}

\item{\code{DiffStepMax}}{Optional. A numeric vector of length equal to number of variables in data containing the maximum acceptable absolute difference between sequential data points for each variable. Defaults to observed maximum (no flags will result)}

\item{\code{DiffPersMin}}{Optional. A numeric vector of length equal to number of variables in data containing the minimum absolute change in value over the interval specified in TintPers to indicate the sensor is not "stuck". Defaults to a vector of zeros (no flags will result).}

\item{\code{TintPers}}{Optional. A difftime object of length equal to number of variables in data specifying the time interval for each variable over which to test for the minimum absolute change in value specified in DiffPersMin. Defaults to 60 x median observed time difference. Class difftime can be generated using as.difftime.}

\item{\code{TestNull}}{Optional. Apply the null test? A logical vector of [TRUE or FALSE] of length equal to number of variables in data. Defaults to FALSE (no null values are flagged)}

\item{\code{NumGap}}{Optional.  A numeric value >= 1, interpretable as an integer, specifying the numer of consecutive NA values constituting a gap. Default is the one more than the length of the data series (no gaps will be flagged)}
}
\value{
A list of flags giving the failed and NA positions for each the Range, Step, Persistence, Null, and Gap tests. Each flag is itself a nested list of failed and na (unable to eval) flagged indices for each variable in data.
}
\description{
Function definition. Determines inplausible data indices based on user-specified limits for the data range, step between adjacent values, persistence (similarity of adjacent values), nulls, and gaps.
}
\examples{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org}
}
\references{
NEON Algorithm Theoretical Basis Document QA/QC Plausibility Testing (NEON.DOC.011081)
license: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{NEON}
\keyword{QAQC,}
\keyword{gap}
\keyword{null,}
\keyword{persistence,}
\keyword{plausibility,}
\keyword{range,}
\keyword{step,}

