% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.amrs.R
\name{def.qf.amrs}
\alias{def.qf.amrs}
\title{Definition function: AMRS flags for XSens}
\usage{
def.qf.amrs(diag32, MethQf = c("qfqm", "xsen")[1])
}
\arguments{
\item{diag32}{The 32-bit diagnostic stream that is output from the XSens AMRS.}

\item{MethQf}{Switch for quality flag determination for the XSens AMRS, diag32 provides ones for passing quality by default the equals "xsen". The "qfqm" method changes the ones to zeros to match the NEON QFQM logic for certain flags that are set high when good, qfAmrsVal & qfAmrsFilt.}
}
\value{
A dataframe (\code{qfSoniAmrs}) of sensor specific AMRS quality flags as described in NEON.DOC.000807.
}
\description{
Definition function to interpret the AMRS (Attitude and Reference Motion System) sensor flags from \code{diag32}. Flags output are a reduced set that were deemed important for the NEON QFQM framework and described in the NEON.DOC.000807.
}
\examples{
diag32 <- as.integer(rep(135, 72000))
def.qf.amrs(diag32 = diag32)

set <- runif(15,1, 72000) # inserting error positions for other flags
diag32[set] <- as.integer(c(262279,524423, 1048710)) # filling with numbers that would indicate flags for qfAmrsVal, qfAmrsVelo, and qfAmrsRng
eddy4R.qaqc::def.qf.amrs(diag32 = diag32)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000807)
XSens AMRS reference manual
}
\seealso{
Currently none
}
\author{
Dave Durden \email{ddurden@battelleecology.org}
}
\keyword{AMRS,}
\keyword{NEON,}
\keyword{qfqm}
