% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.finl.R
\name{def.qf.finl}
\alias{def.qf.finl}
\title{Final Quality Flag (basic L1 data products)}
\usage{
def.qf.finl(data, rtioAlphBeta = c(2, 1), thsh = 20)
}
\arguments{
\item{\code{data}}{A dataframe containing the input flag data that going to use to determine alpha and beta quality flags, quality metrics, and final quality flag. Of class integer". [-]}

\item{\code{rtioAlphBeta}}{Ratio of alpha quailty metric to beta quailty metric for determing the final quality flag. Default to rtioAlphBeta=c(2,1). [-]}

\item{\code{thsh}}{Threshold for determine the condition (pass = 0 or failed = 1) of final quality flag. Default to 20 percent. [percent]}
}
\value{
A list of: \cr
\code{qaqcRpt} A dataframe containing alpha and beta quality flags at the same frequency as input data. [-] \cr
\code{qfqm} A dataframe containing alpha and beta quality metric and final qualiy flag. [-] \cr
}
\description{
Function definition. Determine the final quality flag for individual level 1 data product following the method described in Smith et.al. (2014). The alpha and beta quality flags and metrics were also computed in this function.
}
\examples{
qfA <- c(0,0,0,0,0,1,1,1,1,1,0,0,0,0,1) 
qfB <- c(0,0,-1,-1,-1,1,1,0,0,0,0,0,0,0,0)
qfC <- c(0,1,1,0,0,0,0,0,0,0,0,0,-1,-1,-1)
test<-list()
test$qf <- data.frame(qfA,qfB,qfC)
out <- def.qf.finl(data=test$qf, rtioAlphBeta=c(2,1), thsh=20)
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org} \cr
Natchaya Pingintha-Durden \email{ndurdent@neoninc.org} \cr
}
\references{
NEON Algorithm Theoretical Basis Document: Quality Flags and Quality Metrics for TIS Data Products (NEON.DOC.001113) \cr
Smith, D.E., Metzger, S., and Taylor, J.R.: A transparent and transferable framework for tracking quality information in large datasets. PLoS ONE, 9(11), e112249.doi:10.1371/journal.pone.0112249, 2014. \cr
}
\seealso{
\code{\link[eddy4R.qaqc]{def.qfqm.dp01}} \cr
\code{\link[eddy4R.qaqc]{def.qm}} \cr
}
\keyword{NEON}
\keyword{QAQC,}
\keyword{and}
\keyword{final}
\keyword{flag,}
\keyword{flags}
\keyword{metrics}
\keyword{quality}

