% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.irga.agc.R
\name{def.qf.irga.agc}
\alias{def.qf.irga.agc}
\title{Definition function: Signal strength flag for IRGA}
\usage{
def.qf.irga.agc(qfIrgaAgc, critThsh = 0.5)
}
\arguments{
\item{qfIrgaAgc}{The quality metric derived from the IRGA diagnostics to determine signal strength. Presented as a dimensionless fraction. [-]}

\item{critThsh}{The critical threshold value for the \code{qfIrgaAgc} value to throw the flag for low signal strength (defaults to 0.50 or 50 percent).}
}
\value{
A vector class of integer (\code{qfIrgaAgcOut}) of IRGA AGC flags. Flag indicating when the sensor is operating with low signal strength using 50 percent as the default threshold (1 = when qfIrgaAgc <= 0.50, 0 = when qfIrgaAgc >= 0.50, -1 = NA). [-]
}
\description{
Definition function to generate the signal strength flags for the IRGA from the diagnostic output quality metric \code{qfIrgaAgc}. Flag indicating when the sensor is operating with low signal strength using 50 percent as the default threshold (1 = when qfIrgaAgc <= 0.50, 0 = when qfIrgaAgc >= 0.50, -1 = NA).
}
\examples{
qfIrgaAgc <- rnorm(10,0.70,0.15)
qfIrgaAgc[c(2,8)] <- NA
attributes(qfIrgaAgc)$unit <- "-"

qfIrgaAgcOut <- def.qf.irga.agc(qfIrgaAgc = qfIrgaAgc)

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{IRGA,}
\keyword{NEON,}
\keyword{qfqm}
