% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.irga.vali.R
\name{def.qf.irga.vali}
\alias{def.qf.irga.vali}
\title{Definition function: Validation flag for IRGA}
\usage{
def.qf.irga.vali(frtSet00)
}
\arguments{
\item{frtSet00}{The flow rate set point measured by the IRGA sampling mass flow controller. [dm3 s-1]}
}
\value{
A vector class of integer (\code{qfIrgaVali}) of IRGA validation flags. Flag indicating when the sensor is operated under validation period (1 = validation period, 0 = normal operating condition, -1 = NA) [-]
}
\description{
Definition function to generate the validation flags for IRGA from the IRGA sampling mass flow controller flow rate set point \code{frtSet00}. Flag indicating when the sensor is operated under validation period (1 = validation period, 0 = normal operating condition, -1 = NA).
}
\examples{
qfIrgaVali <- def.qf.irga.vali(frtSet00 = 0.000241)
data <- data.frame(frtSet00 = c(0, 0, 0, -1, -1, -1, 0.000242, 0.000242))
data$qfIrgaVali <- def.qf.irga.vali(frtSet00 = data$frtSet00)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
David Durden \email{ddurden@battelleecology.org}
}
\keyword{IRGA,}
\keyword{NEON,}
\keyword{qfqm}
