% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.rmv.data.R
\name{def.qf.rmv.data}
\alias{def.qf.rmv.data}
\title{Definition function: to remove high frequency data points that have failed quality flags}
\usage{
def.qf.rmv.data(dfData, dfQf, Sens = NULL, Vrbs = FALSE,
  TypeData = c("integer", "real")[1])
}
\arguments{
\item{dfData}{Input data.frame for data to be removed from based on quality flags}

\item{dfQf}{Input data.frame of quality flags (must be of class integer to be included in the processing)}

\item{Sens}{Character string indicating the sensor the high frequency data come from to check for sensor specific flags}

\item{Vrbs}{Optional. A logical {FALSE/TRUE} value indicating whether to:\cr
\code{Vrbs = FALSE}: (Default) cleaned data set, list of variables assessed, list of quality flags for each variable assessed, and the total number of bad data per variable, or \cr
\code{Vrbs = TRUE}: cleaned data set, list of variables assessed, list of quality flags for each variable assessed, the number of each quality flag tripped for each variable and the total number of bad data per variable}

\item{TypeData}{Type of input data, TypeData = c("integer", "real"). Defaults to "integer". [-].}
}
\value{
A list (\code{rpt}) containing a dataframe (\code{rpt$dfData}) of the data with bad data replaced by NaN's, a vector of data variables to assess (\code{rpt$listVar}),  a list containing vectors of flag names used for each variable (\code{rpt$listQf}),  a list containing vectors of the number of quality flags set high for each variable (\code{rpt$fracQfBad}), a list containing vectors of flagged data points for each variable (\code{rpt$posBad}), a list containing total number of flagged data points for each variable (\code{rpt$numBadSum}).
}
\description{
Definition function  to remove high frequency data points that have failed quality flags from a data.frame
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000807) \cr
}
\seealso{
Currently none
}
\author{
Dave Durden \email{ddurden@battelleecology.org}
}
\keyword{NEON,}
\keyword{qfqm,}
\keyword{quality}
