% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.shdw.R
\name{def.qf.shdw}
\alias{def.qf.shdw}
\title{Definition function: Tower shadowing flags for CSAT3}
\usage{
def.qf.shdw(angWindMean, Site, Crs = c("NED", "ENU")[1])
}
\arguments{
\item{angWindMean}{The mean wind direction from the Soni (AngZaxsErth) in radians.}

\item{Site}{The NEON 4-letter site code.}
}
\value{
A dataframe (\code{qfShdw}) of sonic anemometer qfShdw quality flag.
}
\description{
Definition function to diagnose shadow effects due to tower structure on the sonic anemometer data
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0' data product conversions and calculations (NEON.DOC.000807) \cr
Campbell Scientific CSAT3 reference manual
}
\seealso{
Currently none
}
\author{
Sreenath Paleri \email{paleri@battelleecology.org}
}
\keyword{CSAT3,}
\keyword{NEON,}
\keyword{anemometer,}
\keyword{qfqm,}
\keyword{shadow}
\keyword{soni,}
\keyword{sonic}
\keyword{tower}
