% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.soni.R
\name{def.qf.soni}
\alias{def.qf.soni}
\title{Definition function: Soni flags for CSAT3}
\usage{
def.qf.soni(diag16)
}
\arguments{
\item{diag16}{The 16-bit diagnostic stream that is output from the Soni (CSAT3).}
}
\value{
A dataframe (\code{qfSoni}) of sensor specific AMRS quality flags as described in NEON.DOC.000807.
}
\description{
Definition function to interpret the Soni (Campbell Sci. CSAT3) sensor flags from \code{diag16}. Flags output are important for the NEON QFQM framework and described in the NEON.DOC.000807. This code was built for flags outlined in firmware version 3 of the CSAT3; however, flag detemination may be transferable to other firmware versions.
}
\examples{
diag16 <- as.integer(rep(135, 36000))
pos <- runif(20,1, 36000) # inserting error positions for other flags
diag16[pos] <- as.integer(c(32768,16384,8192,4096,61442, 61441,61440,61503, -99999, NaN)) # filling with numbers that would indicate flags soni flags

eddy4R.qaqc::def.qf.soni(diag16 = diag16)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000807) \cr
Campbell Scientific CSAT3 reference manual
}
\seealso{
Currently none
}
\author{
Dave Durden \email{ddurden@battelleecology.org}
}
\keyword{CSAT3,}
\keyword{NEON,}
\keyword{anemometer,}
\keyword{qfqm}
\keyword{soni,}
\keyword{sonic}
