% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qfqm.dp01.R
\name{def.qfqm.dp01}
\alias{def.qfqm.dp01}
\title{Quality flags and quality metrics (basic L1 data products)}
\usage{
def.qfqm.dp01(data, time = base::as.POSIXlt(base::seq.POSIXt(from =
  base::Sys.time(), by = "sec", length.out = base::length(data[, 1]))),
  posQf = base::vector("list", base::length(data)),
  qf = base::vector("list", base::length(data)), WndwAgr = 1800 *
  stats::median(base::abs(base::diff(time)), na.rm = TRUE),
  NameQfExcl = base::as.list(base::character(length = base::length(data))))
}
\arguments{
\item{data}{Required input. A data frame containing the L0 data evaluated (do not include the time stamp vector here).}

\item{time}{Optional. A time vector of class POSIXlt of times corresponding with each row in data. Defaults to an evenly spaced time vector starting from system time of execution by seconds.}

\item{posQf}{Optional. Only input ONE of either \code{posQf} or \code{qf}. # A named list of variables matching those in data, each itself a named list of flags (standard or user-defined names), with nested lists of $fail and $na vector positions of failed and na quality tests for that variable and flag (eg. posQf$X$posQfStep$fail and posQf$Y$posQfStep$na). Aggregated quality flags and metrics will be computed for all flags in this list, and all will be used to compute the final quality flag. This function directly accepts the output from def.plaus.R and def.dspk.wndw.R. Defaults to an empty flag list.}

\item{qf}{Optional. Only input ONE of either \code{posQf} or \code{qf}. A list of variables matching those in data, each containing a data frame of quality flags for that variable. Number of rows must match that of \code{data}}

\item{WndwAgr}{Optional. A difftime object of length equal to number of variables in data specifying the time interval for aggregating flags of each variable. Defaults to 1800 x median observed time difference. Class difftime can be generated using as.difftime.}

\item{NameQfExcl}{Optional. A list of length equal to number of variables, each itself a vector of strings naming the flags (matching a subset of posQf) for which to exclude the flagged indices of each variable from the L1 average. Eg. NameQfExcl <- list(x=c("posQfRng","posQfStep"),y=c("posQfPers","posQfNull","posQfGap")). Note that variables in the list must be in the same order as those in data. Defaults to an empty list (flagged points are not excluded from the L1 average).}
}
\value{
A list of: \cr
timeAgrBgn - the starting time stamp of aggregated L1 data and quality metrics \cr
dataAgr - a list of variables, each containing a data frame of the time-aggregated mean, minimum, maximum, variance, standard deviation of the mean, number of points going into the average, and quality metrics (pass, fail, NA) pertaining to that variable for each flag in posQf, as well as the alpha & beta quality metrics and final quality flag.
}
\description{
Function definition. Aggregates the quality flags and computes quality metrics and final quality flag for basic L1 (time window averaged) data products from the plausibility of the individual (L0) data indices to be aggregated.
}
\examples{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org}
}
\references{
NEON Algorithm Theoretical Basis Document: Quality Flags and Quality Metrics for TIS Data Products (NEON.DOC.001113)
}
\seealso{
Currently none
}
\keyword{L1}
\keyword{NEON}
\keyword{QAQC,}
\keyword{and}
\keyword{average,}
\keyword{final}
\keyword{flag}
\keyword{flags}
\keyword{metrics,}
\keyword{quality}

