% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qfqm.l1.R
\name{def.qfqm.l1}
\alias{def.qfqm.l1}
\title{Quality flags and quality metrics (basic L1 data products)}
\usage{
def.qfqm.l1(data, ts = as.POSIXlt(seq.POSIXt(from = Sys.time(), by = "sec",
  length.out = length(data[, 1]))), flgs = list(), TintAggr = 1800 *
  median(abs(diff(ts)), na.rm = TRUE), FlagExcl = as.list(character(length =
  length(data))))
}
\arguments{
\item{data}{Required input. A data frame containing the L0 data evaluated (do not include the time stamp vector here).}

\item{ts}{Optional. A time vector of class POSIXlt of times corresponding with each row in data. Defaults to an evenly spaced time vector starting from system time of execution by seconds.}

\item{flgs}{Optional. A list of named flags (standard or user-defined or names), each itself a list of flagged indices for each variable in data (ex. flgs$posFlagStep$X and flgs$posFlagStep$Y listing flagged positions for variables X and Y, respectively). Aggregated quality flags and metrics will be computed for all flags in this list, and all will be used to compute the final quality flag. This function directly accepts the output from def.plaus.R and def.dspk.wndw.R. Defaults to an empty flag list.}

\item{TintAggr}{Optional. A difftime object of length equal to number of variables in data specifying the time interval for aggregating flags of each variable. Defaults to 1800 x median observed time difference. Class difftime can be generated using as.difftime.}

\item{FlagExcl}{Optional. A list of length equal to number of variables, each itself a vector of strings naming the flags (matching a subset of flgs) for which to exclude the flagged indices of each variable from the L1 average. Eg. FlagExcl <- list(x=c("posFlagRng","posFlagStep"),y=c("posFlagPers","posFlagNull","posFlagGap")). Note that variables in the list must be in the same order as those in data. Defaults to an empty list (flagged points are not excluded from the L1 average).}
}
\value{
A list of: \cr
tsAggrBgn - the starting time stamp of aggregated L1 data and quality metrics \cr
qfqm - a list of variables, each containing a data frame of the time-aggregated mean, minimum, maximum, variance, standard deviation of the mean, number of points going into the average, and quality metrics (pass, fail, NA) pertaining to that variable for each flag in flgs, as well as the alpha & beta quality metrics and final quality flag.
}
\description{
Function definition. Aggregates the quality flags and computes quality metrics and final quality flag for basic L1 (time window averaged) data products from the plausibility of the individual (L0) data indices to be aggregated.
}
\examples{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org}
}
\references{
NEON Algorithm Theoretical Basis Document: Quality Flags and Quality Metrics for TIS Data Products (NEON.DOC.001113)
}
\seealso{
Currently none
}
\keyword{L1}
\keyword{NEON}
\keyword{QAQC,}
\keyword{and}
\keyword{average,}
\keyword{final}
\keyword{flag}
\keyword{flags}
\keyword{metrics,}
\keyword{quality}

