% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qm.R
\name{def.qm}
\alias{def.qm}
\title{Quality Metrics}
\usage{
def.qm(data, nameFlag)
}
\arguments{
\item{\code{data}}{A dataframe containing the input flag data. Of class integer". [-]}

\item{\code{nameFlag}}{A vector of class "character" containing the name of flag data which will be used to calculate quality metrics of failed, pass, NA, alpha, and beta and final quality flag. [-]}
}
\value{
A dataframe containing quality metrics of failed, pass, and NA for each of the individual flag defined in \code{nameFlag}.
}
\description{
Function definition. Determine the quality metrics of failed, pass, and NA for each of the individual quality flag following the method described in Smith et.al. (2014).
}
\examples{
qfA <- c(0,0,0,0,0,1,1,1,1,1,0,0,0,0,1) 
qfB <- c(0,0,-1,-1,-1,1,1,0,0,0,0,0,0,0,0)
qfC <- c(0,1,1,0,0,0,0,0,0,0,0,0,-1,-1,-1)
test<-list()
test$qf <- data.frame(qfA,qfB,qfC)
test$qm<- def.qm(data=test$qf, nameFlag=colnames(test$qf))
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org} \cr
Natchaya Pingintha-Durden \email{ndurdent@neoninc.org} \cr
}
\references{
NEON Algorithm Theoretical Basis Document: Quality Flags and Quality Metrics for TIS Data Products (NEON.DOC.001113) \cr
Smith, D.E., Metzger, S., and Taylor, J.R.: A transparent and transferable framework for tracking quality information in large datasets. PLoS ONE, 9(11), e112249.doi:10.1371/journal.pone.0112249, 2014. \cr
}
\seealso{
\code{\link[eddy4R.qaqc]{def.qfqm.l1}}
}
\keyword{NEON}
\keyword{QAQC,}
\keyword{flags,}
\keyword{metrics}
\keyword{quality}

