% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.neon.dp01.qfqm.ecse.R
\name{wrap.neon.dp01.qfqm.ecse}
\alias{wrap.neon.dp01.qfqm.ecse}
\title{Wrapper function:  Preprocessing and calculating quality metrics, alpha and beta quality metrics, and final quality flag for the NEON eddy-covariance stroage exchange L1 data products}
\usage{
wrap.neon.dp01.qfqm.ecse(dp01 = c("co2Stor", "h2oStor", "tempAirLvl",
  "tempAirTop", "isoCo2", "isoH2o")[1], lvl, lvlMfcSampStor = NULL,
  valvLvl = NULL, lvlCrdH2oValvVali = NULL, data = list(),
  qfInput = list(), TypeMeas = c("samp", "vali")[1], PrdMeas, PrdAgr,
  idxTime = list())
}
\arguments{
\item{\code{dp01}}{A vector of class "character" containing the name of NEON ECSE dp01 which descriptive statistics are being calculated, \cr
c("co2Stor", "h2oStor", "tempAirLvl", "tempAirTop", "isoCo2", "isoH2o"). Defaults to "co2Stor". [-]}

\item{\code{lvl}}{Measurement level of dp01 which descriptive statistics are being calculated. Of type character. [-]}

\item{\code{lvlMfcSampStor}}{Measurement level of mfcSampStor which apply to only  dp01 equal to "co2Stor" or "h2oStor". Defaults to NULL. Of type character. [-]}

\item{\code{valvLvl}}{Measurement level of irgaValvLvl, crdCo2ValvLvl, or crdH2oValvLvl. Defaults to NULL. Of type character. [-]}

\item{\code{lvlCrdH2oValvVali}}{Measurement level of crdH2oValvVali which apply to only  dp01 equal to "isoH2o". Defaults to NULL. Of type character. [-]}

\item{\code{data}}{A list of data frame containing the input dp0p data that related to dp01 which qfqm are being calculated. Of class integer". [User defined]}

\item{\code{qfInput}}{A list of data frame containing the input quality flag data that related to dp01 are being grouped. Of class integer". [NA]}

\item{\code{TypeMeas}}{A vector of class "character" containing the name of measurement type (sampling or validation), TypeMeas = c("samp", "vali"). Defaults to "samp". [-]}

\item{\code{PrdMeas}}{The measurement time period in minute.  [min]}

\item{\code{PrdAgr}}{The time period to aggregate to averaging in minute. [min]}

\item{\code{idxTime}}{A list of data frame containing the indices and corresponding times for aggregation periods. [-]}
}
\value{
A list of dp01 descriptive statistics. \cr 
\code{mean} The mean of non-NA values in \code{data}
\code{min} The minimum value of non-NA values in \code{data}
\code{max} The maximum value of non-NA values in \code{data}
\code{vari} The variance of non-NA values in \code{data}
\code{numSamp} The number of non-NA values in \code{data}
\code{se} The standard error of the mean of non-NA values in \code{data}
\code{timeBgn} The beginning time to determine descriptive statistics.
\code{timeEnd} The ending time to determine descriptive statistics.
}
\description{
Wrapper function. Preprocessing and calculating quality metrics, alpha and beta quality metrics, and final quality flag for the NEON eddy-covariance stroage exchange (ECSE) Level 1 data products (dp01).
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Natchaya Pingintha-Durden \email{eddy4R.info@gmail.com}
}
\keyword{ECSE,}
\keyword{NEON}
\keyword{QAQC,}
\keyword{dp01,}
\keyword{flags,}
\keyword{metrics,}
\keyword{quality}
