% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.qf.rmv.data.R
\name{wrap.qf.rmv.data}
\alias{wrap.qf.rmv.data}
\title{Wrapper function: to remove high frequency data points that have failed quality flags}
\usage{
wrap.qf.rmv.data(
  inpList,
  Sens = NULL,
  DspkSensRmv = c("amrs", NA)[1],
  qfRmv = NULL,
  Vrbs = FALSE,
  MethMeas = c("ecte", "ecse")[1],
  MethFf = TRUE
)
}
\arguments{
\item{inpList}{List consisting of \code{ff::ffdf} file-backed objects, in the format provided by function \code{eddy4R.base::wrap.hdf5.read()}. Of types numeric and integer.}

\item{Sens}{Character string indicating which sensor will remove high frequency data points that have failed quality flags. Defaults to NULL.#'}

\item{DspkSensRmv}{Character string indicating which sensor will be removed from despiking algorithm. Defaults to "amrs".}

\item{qfRmv}{Character string indicating which quality flag will exclude in the processing. Defaults to NULL.}

\item{Vrbs}{Optional. A logical {FALSE/TRUE} value indicating whether to:\cr
\code{Vrbs = FALSE}: (Default) cleaned data set with the bad high frequency quality flagged data replaced with NaN's as part of the \code{inpList} in the same format., or \cr
\code{Vrbs = TRUE}: cleaned data set with the bad high frequency quality flagged data replaced with NaN's as part of the \code{inpList} in the same format. In addition, a separate list  \code{qfqmAnls} will be added to the output list \code{rpt} with a list of variables assessed, a list of quality flags for each variable assessed, the number of each quality flag tripped for each variable and the total number of bad data per variable.}

\item{MethMeas}{A vector of class "character" containing the name of measurement method (eddy-covariance turbulent exchange or storage exchange), MethMeas = c("ecte", "ecse"). Defaults to "ecte". [-]}

\item{MethFf}{A logical {FALSE/TRUE} value indicating whether to output as ff object or regular data.frame. Defaults to \code{MethFf = TRUE}}
}
\value{
The returned object consistes of \code{inpList}, with the bad high frequency quality flagged data replaced with NaN's. Optionally, (\code{Vrbs = TRUE}) a separate list \code{qfqmAnls} will be added to the output list \code{rpt} with a list of variables assessed, a list of quality flags for each variable assessed, the number of each quality flag tripped for each variable and the total number of bad data per variable.
}
\description{
Wrapper function  to remove high frequency data points that have failed quality flags from a data.frame
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0' data product conversions and calculations (NEON.DOC.000807) \cr
Licor LI7200 reference manual
}
\seealso{
Currently none
}
\author{
Dave Durden \email{ddurden@battelleecology.org}
}
\keyword{NEON,}
\keyword{qfqm,}
\keyword{quality}
