% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.qfqm.rpt.R
\name{wrap.qfqm.rpt}
\alias{wrap.qfqm.rpt}
\title{Wrapper function: QFQM quality report for eddy-covariance dp04 HDF5 files}
\usage{
wrap.qfqm.rpt(
  FileName,
  Dp = NULL,
  Vrbs = FALSE,
  MethQm = c("Pass", "Fail", "Na")[1]
)
}
\arguments{
\item{Dp}{A vector of data products to check the quality report for. Defaults to NULL which will test all dp's found in HDF5 file. Of type character. [-]}

\item{Vrbs}{A logical parameter that determines if all failed quality times should be reported (defaults to FALSE).}

\item{MethQm}{A character string that determines which quality metric to pull ("Pass", "Fail", "Na").}

\item{File}{character vector for the full path to the HDF5 file to perform the quality report [-]}
}
\value{
A list for all dp's provided including either summary or verbose quality report: \cr
\code{qm}   quality metrics (percent) passed sub-data products if Vrbs = TRUE. [percent] \cr
\code{qmAlph} In dataframe column as "numeric" containing the alpha quality metric for sub-data products. [percent] \cr
\code{qmBeta} In dataframe column as "numeric" containing the beta quality metric for sub-data products. [percent] \cr
\code{qfFinl} In dataframe column as class "integer", [0,1], containing the final quality flag for sub-data products. [-] \cr
\code{qfSciRevw} In dataframe column as class "integer", [0,1], containing the scientific review quality flag for sub-data products. [-] \cr
}
\description{
Wrapper function. Summarize the qfqm reports for NEON expanded eddy-covariance HDF5 files.
}
\examples{
#no example at this time
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{NEON}
\keyword{QAQC,}
\keyword{flags,}
\keyword{metrics}
\keyword{quality}
