% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.agr.file.dp00.R
\name{def.agr.file.dp00}
\alias{def.agr.file.dp00}
\title{Definition function: Combine individual L0 data streams from file}
\usage{
def.agr.file.dp00(
  dirFile = ".",
  nameFile,
  nameVar,
  unitVar,
  Freq,
  FmtTime = "\%d-\%b-\%Y \%I.\%M.\%OS \%p",
  Tz = "GMT"
)
}
\arguments{
\item{dirFile}{Optional. A character string indicating the directory in which the files to combine are stored. Default is the current working directory.}

\item{nameFile}{Required. A character vector of file names holding the L0 data streams to combine. Must be in .csv format. Within these files, the first column must containing the measurement time, the second column must contain one L0 data stream}

\item{nameVar}{Required. A character vector of the same length as nameFile with the variable names of the L0 data streams}

\item{unitVar}{Required. A character vector of the same length as nameFile with the variable units}

\item{Freq}{Required. A numeric value indicating the expected frequency [Hz] of L0 data within the files in nameFile}

\item{FmtTime}{Optional. An character format string to interpret the time values in the first column of each file in nameFile. Default is \%d-\%b-\%Y \%I.\%M.\%OS \%p}

\item{Tz}{Optional. A character string specifying the time zone in with the time values in the first column of each file in nameFile are represented. Default is GMT}
}
\value{
A list of: \cr
\code{time} a POSIXlt vector of regularized times corresponding to each row in data. Limits are the min and max of times found within nameFile \cr
\code{data} a named data frame containing the regularized time series of L0 variables found within nameFile
}
\description{
Definition function. Combine (and regularize) L0 data streams downloaded in individual files from the L0 sandbox tool.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\keyword{L0}
\keyword{NEON,}
