% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.plau.R
\name{def.plau}
\alias{def.plau}
\title{Definition function: Plausibility tests (Range, Step, Persistence, Null, Gap)}
\usage{
def.plau(
  data,
  time = as.POSIXlt(seq.POSIXt(from = Sys.time(), by = "sec", length.out = nrow(data))),
  RngMin = apply(data, 2, min, na.rm = TRUE),
  RngMax = apply(data, 2, max, na.rm = TRUE),
  DiffStepMax = apply(abs(apply(data, 2, diff)), 2, max, na.rm = TRUE),
  DiffPersMin = rep.int(0, ncol(data)),
  WndwPers = 60 * median(abs(diff(time)), na.rm = TRUE) * rep.int(1, ncol(data)),
  TestNull = rep(FALSE, ncol(data)),
  NumGap = rep(nrow(data) + 1, ncol(data)),
  Vrbs = FALSE
)
}
\arguments{
\item{data}{Required input. A data frame containing the data to be evaluated (do not include the time stamp vector here).}

\item{time}{Optional. A time vector of class POSIXlt of times corresponding with each row in data. Defaults to an evenly spaced time vector starting from system time of execution by seconds.}

\item{RngMin}{Optional. A numeric vector of length equal to number of variables in data containing the minimum acceptable value for each variable. Defaults to observed minimums (no flags will result)}

\item{RngMax}{Optional. A numeric vector of length equal to number of variables in data containing the maximum acceptable value for each variable. Defaults to observed maximums (no flags will result)}

\item{DiffStepMax}{Optional. A numeric vector of length equal to number of variables in data containing the maximum acceptable absolute difference between sequential data points for each variable. Defaults to observed maximum (no flags will result)}

\item{DiffPersMin}{Optional. A numeric vector of length equal to number of variables in data containing the minimum absolute change in value over the interval specified in WndwPers to indicate the sensor is not "stuck". Defaults to a vector of zeros (no flags will result).}

\item{WndwPers}{Optional. The time window for evaluting the persistence test. This must be a vector of length equal to number of variables in data. If the values are numeric (integer), then WndwPers specifies number of data points over which to test for the minimum absolute change in value specified in DiffPersMin. 
If the vector is a difftime object (e.g. as.difftime(5,units="secs")), it specifies the time interval over which to test for the minimum absolute change in value specified in DiffPersMin. The results are the same if the time-based window exactly corresponds to an integer number of data points.
Defaults to a difftime object of 60 x median observed time difference.}

\item{TestNull}{Optional. Apply the null test? A logical vector of [TRUE or FALSE] of length equal to number of variables in data. Defaults to FALSE (no null values are flagged)}

\item{NumGap}{Optional.  A numeric value >= 1, interpretable as an integer, specifying the numer of consecutive NA values constituting a gap. Default is the one more than the length of the data series (no gaps will be flagged)}

\item{Vrbs}{Optional. A logical {FALSE/TRUE} value indicating whether to:\cr
\code{Vrbs = FALSE}: (Default) output the vector positions of the fail and na results for each test (default), or \cr
\code{Vrbs = TRUE}: output a data frame for each variable in data, with a column for each plausibility test outputting the actual quality flags [-1,0,1] for each data point}
}
\value{
If:\cr
\code{Vrbs = FALSE} A list of variables in \code{data}, nested within each a list of qf test names: (\code{setQfRng}), Step (\code{setQfStep}), Persistence  (\code{setQfPers}), Null (\code{setQfNull}), and Gap(\code{setQfGap}) tests.
Each flag is itself a nested list of failed and na (unable to eval) flagged vector positions (e.g. \code{$X$setQfRng$fail} and \code{$X$setQfRng$na} \cr
\code{Vrbs = TRUE} A list of variables matching those in \code{data}, each containing a data frame with a column for each plausibility test outputting the actual quality flags [-1,0,1] for each data point, where -1 indicates the test could not be evaluated, 0 indicates a pass, and 1 indicates a fail
}
\description{
Function definition. Determines inplausible data indices based on user-specified limits for the data range, step between adjacent values, persistence (similarity of adjacent values), nulls, and gaps.
}
\examples{
data <- data.frame(x=rnorm(1000,mean=0,sd=1)) # Start off with a vector of 1000 random values
data$x[c(20,50,500,90)] <- 50 # insert some spikes
data$x[600:699] <- rnorm(100,mean=0,sd=0.001) # Add some "stuck" data
data$x[800:810] <- NA
data$x[400:450] <- Inf # Results in Error
RngMin <- -4
RngMax <- 4
DiffStepMax <- 6
DiffPersMin <- 0.1
WndwPers <- as.difftime(10,units="secs") # We are using the default time variable, which generates a freq of 1 second
TestNull <- TRUE
NumGap <- 11
Vrbs=TRUE
qf <- eddy4R.qaqc::def.plau(data=data,RngMin=RngMin,RngMax=RngMax,DiffStepMax=DiffStepMax,DiffPersMin=DiffPersMin,WndwPers=WndwPers,TestNull=TestNull,NumGap=NumGap,Vrbs=Vrbs)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007 \cr
NEON Algorithm Theoretical Basis Document QA/QC Plausibility Testing (NEON.DOC.011081)
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\keyword{NEON}
\keyword{QAQC,}
\keyword{gap}
\keyword{null,}
\keyword{persistence,}
\keyword{plausibility,}
\keyword{range,}
\keyword{step,}
