% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qf.irga.R
\name{def.qf.irga}
\alias{def.qf.irga}
\title{Definition function: irga flags for LI72000}
\usage{
def.qf.irga(diag01, MethQf = c("qfqm", "lico")[1])
}
\arguments{
\item{diag01}{The 32-bit diagnostic stream that is output from the LI7200.}

\item{MethQf}{Switch for quality flag determination for the LI7200, diag01 provides ones for passing quality by default the equals "lico". The "qfqm" method changes the ones to zeros to match the NEON QFQM logic.}
}
\value{
A dataframe (\code{qfIrgaTurb}) of sensor specific irga quality flags as described in NEON.DOC.000807.
}
\description{
Definition function to interpret the irga sensor flags from \code{diag01}.
}
\examples{
diag01 <- rep(8191, 72000)
def.qf.irga(diag01 = diag01)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0' data product conversions and calculations (NEON.DOC.000807) \cr
Licor LI7200 reference manual
}
\seealso{
Currently none
}
\author{
Dave Durden \email{ddurden@battelleecology.org}
}
\keyword{NEON,}
\keyword{irga,}
\keyword{qfqm}
