% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.qm.R
\name{def.qm}
\alias{def.qm}
\title{Quality Metrics}
\usage{
def.qm(qf, nameQmOut = NULL)
}
\arguments{
\item{qf}{A data frame of quality flags, class integer. Each column contains the quality flag values [-1,0,1] for that flag. Note: This is the Vrbs output from def.plau, def.dspk.wndw, and def.dspk.filt.med. See def.conv.qf.vrbs for converting from non-verbose to verbose output.}

\item{nameQmOut}{Optional. A vector of class "character" containing the base name of the output quality metrics for each flag in \code{qf}. These names will be ammended with "Pass", "Fail", or "Na" at the end when outputting their respective quality metrics. Default behavoir is to autoassign names based on the column names of \code{qf} [-]}
}
\value{
A dataframe containing quality metrics (fractions) of failed, pass, and NA for each of the individual flag defined in \code{qf}.
}
\description{
Function definition. Determine the quality metrics of failed, pass, and NA for each of the individual quality flag following the method described in Smith et.al. (2014). Performed for the entire set of input data.
}
\examples{
qfA <- c(0,0,0,0,0,1,1,1,1,1,0,0,0,0,1) 
qfB <- c(0,0,-1,-1,-1,1,1,0,0,0,0,0,0,0,0)
qfC <- c(0,1,1,0,0,0,0,0,0,0,0,0,-1,-1,-1)
test<-list()
test$qf <- data.frame(qfA,qfB,qfC)
test$qm<- def.qm(qf=test$qf, nameQmOut=c("qmA","qmB","qmC"))
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org} \cr
Natchaya Pingintha-Durden \email{ndurdent@neoninc.org} \cr
}
\references{
NEON Algorithm Theoretical Basis Document: Quality Flags and Quality Metrics for TIS Data Products (NEON.DOC.001113) \cr
Smith, D.E., Metzger, S., and Taylor, J.R.: A transparent and transferable framework for tracking quality information in large datasets. PLoS ONE, 9(11), e112249.doi:10.1371/journal.pone.0112249, 2014. \cr
}
\seealso{
\code{\link[eddy4R.qaqc]{wrap.qfqm.dp01}}
\code{\link[eddy4R.qaqc]{def.plau}}
\code{\link[eddy4R.qaqc]{def.dspk.wndw}}
\code{\link[eddy4R.qaqc]{def.dspk.filt.med}}
\code{\link[eddy4R.qaqc]{def.conv.qf.vrbs}}
}
\keyword{NEON}
\keyword{QAQC,}
\keyword{flags,}
\keyword{metrics}
\keyword{quality}

